scripts/requirements.txt
confluent_kafka
types-confluent-kafka
psycopg2-binary
